+++ changeLog.md
@@ -1,3 +1,4 @@
 ## March 23 2024
 * removed extra quit button in system tray icon when minimized. system tray icon now always visible (previously icon only show when main window not visible.).
 * added opacity options in snapshop context menu and hotkeys. 
+* added opacity snapshot opacity change on mouse hover (to see stuff under the snapshot without moving it)